HPゲージバグの修正及び変数、関数のリファクタリング

+++ MyDX12/GameObject/Boss.h
@@ -49,7 +49,7 @@ namespace XIIlib {
 
 		void IniState()override;
 
-		void SetHitDamage(float attackPoint)override;				// ダメージを受ける関数
+		void SetHitDamage(int attackPoint)override;				// ダメージを受ける関数
 
 		void CreateAttackArea()override;
